modules/audio-control/modules/launch-control-xl3/midi-server/requirements.txt
python-rtmidi
flask
flask-cors
flask-socketio
python-socketio
eventlet